requirements.txt
dj-database-url==0.4.2
Django==1.11.4
django-cors-headers==2.1.0
django-filter==1.0.4
djangorestframework==3.6.3
enum34==1.1.6
gunicorn==19.7.1
httpie==0.9.9
idna==2.6
ipython==5.4.1
ipython-genutils==0.2.0
olefile==0.44
pathlib2==2.3.0
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.2.1
prompt-toolkit==1.0.15
psycopg2==2.7.3
ptyprocess==0.5.2
Pygments==2.2.0
python-decouple==3.1
pytz==2017.2
whitenoise==3.3.1
